[REFACTOR] 디렉토리 분리 - 어르신 기본정보 (basic)
