+++ src/main/java/com/balybus/galaxy/patient/dto/PatientRequestDto.java
@@ -4,6 +4,7 @@ import com.balybus.galaxy.global.domain.tblAddressFirst.TblAddressFirst;
 import com.balybus.galaxy.global.domain.tblAddressSecond.TblAddressSecond;
 import com.balybus.galaxy.global.domain.tblAddressThird.TblAddressThird;
 import com.balybus.galaxy.global.domain.tblCenterManager.TblCenterManager;
+import com.balybus.galaxy.patient.basic.dto.BasicRequestDto;
 import com.balybus.galaxy.patient.domain.tblPatient.TblPatient;
 import com.balybus.galaxy.patient.domain.tblPatientLog.TblPatientLog;
 import com.balybus.galaxy.patient.dto.baseDto.PatientBaseDto;
@@ -11,52 +12,52 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 
 public class PatientRequestDto {
-    @Data
-    @EqualsAndHashCode(callSuper = true)
-    public static class SavePatientInfo extends PatientBaseDto {
-        // entity 전환
-        public TblPatient toEntity(
-                TblCenterManager manager, TblAddressFirst tblAddressFirst,
-                TblAddressSecond tblAddressSecond, TblAddressThird tblAddressThird){
-            return TblPatient.builder()
-                    .manager(manager)
-                    .tblAddressFirst(tblAddressFirst)
-                    .tblAddressSecond(tblAddressSecond)
-                    .tblAddressThird(tblAddressThird)
-                    .careLevel(this.careLevel)
-                    .inmateState(this.inmateState)
-                    .workType(this.workType)
-                    .gender(this.gender)
-                    .dementiaSymptom(this.dementiaSymptom)
-                    .serviceMeal(this.serviceMeal)
-                    .serviceToilet(this.serviceToilet)
-                    .serviceMobility(this.serviceMobility)
-                    .serviceDaily(this.serviceDaily)
-                    .name(this.name)
-                    .birthDate(this.birthDate)
-                    .weight(this.weight)
-                    .diseases(this.diseases)
-                    .timeNegotiation(this.timeNegotiation)
-                    .build();
-        }
-    }
+//    @Data
+//    @EqualsAndHashCode(callSuper = true)
+//    public static class SavePatientInfo extends PatientBaseDto {
+//        // entity 전환
+//        public TblPatient toEntity(
+//                TblCenterManager manager, TblAddressFirst tblAddressFirst,
+//                TblAddressSecond tblAddressSecond, TblAddressThird tblAddressThird){
+//            return TblPatient.builder()
+//                    .manager(manager)
+//                    .tblAddressFirst(tblAddressFirst)
+//                    .tblAddressSecond(tblAddressSecond)
+//                    .tblAddressThird(tblAddressThird)
+//                    .careLevel(this.careLevel)
+//                    .inmateState(this.inmateState)
+//                    .workType(this.workType)
+//                    .gender(this.gender)
+//                    .dementiaSymptom(this.dementiaSymptom)
+//                    .serviceMeal(this.serviceMeal)
+//                    .serviceToilet(this.serviceToilet)
+//                    .serviceMobility(this.serviceMobility)
+//                    .serviceDaily(this.serviceDaily)
+//                    .name(this.name)
+//                    .birthDate(this.birthDate)
+//                    .weight(this.weight)
+//                    .diseases(this.diseases)
+//                    .timeNegotiation(this.timeNegotiation)
+//                    .build();
+//        }
+//    }
+
+//    @Data
+//    @EqualsAndHashCode(callSuper = true)
+//    public static class UpdatePatientInfo extends PatientBaseDto {
+//        private Long patientSeq;      //어르신 구분자
+//
+//    }
+
+//    @Data
+//    public static class GetPatientList {
+//        private Integer pageNo;
+//        private Integer pageSize;
+//    }
 
     @Data
     @EqualsAndHashCode(callSuper = true)
-    public static class UpdatePatientInfo extends PatientBaseDto {
-        private Long patientSeq;      //어르신 구분자
-
-    }
-
-    @Data
-    public static class GetPatientList {
-        private Integer pageNo;
-        private Integer pageSize;
-    }
-
-    @Data
-    @EqualsAndHashCode(callSuper = true)
-    public static class RecruitHelper extends UpdatePatientInfo {
+    public static class RecruitHelper extends BasicRequestDto.UpdatePatientInfo {
         private Boolean linkingYn;          //어르신 기본정보로 업데이트 할 것인지 여부
         private Boolean wageNegotiation;    //급여 협의 여부
         private int wageState;              // 1:시급, 2:일급, 3:주급 구분
