+++ src/main/java/com/balybus/galaxy/patient/service/serviceImpl/PatientServiceImpl.java
@@ -95,144 +95,144 @@ public class PatientServiceImpl implements PatientService {
     }
 
 
-    /**
-     * 어르신 정보 등록
-     * @param userEmail String:토큰 조회 결과 사용자 이메일 데이터
-     * @param dto PatientRequestDto.SavePatientInfo
-     * @return PatientResponseDto.SavePatientInfo
-     */
-    @Override
-    @Transactional
-    public PatientResponseDto.SavePatientInfo savePatientInfo(String userEmail, PatientRequestDto.SavePatientInfo dto) {
-        //1. 관리자 정보 조회
-        TblCenterManager centerManager = loginAuthCheckService.checkManager(userEmail);
-
-        //2. 어르신 정보 등록
-        //2-1. 주소 정보 검증
-        TblAddressFirst firstAddress = firstAddressService.validationCheck(dto.getAfSeq());
-        TblAddressSecond secondAddress = secondAddressService.validationCheck(dto.getAfSeq(), dto.getAsSeq());
-        TblAddressThird thirdAddress = thirdAddressService.validationCheck(dto.getAsSeq(), dto.getAtSeq());
-
-        //2-2. 어르신 정보 entity 전환 및 저장
-        TblPatient patient = patientRepository.save(dto.toEntity(centerManager, firstAddress, secondAddress, thirdAddress));
-
-        //3. 어르신 돌봄 시간 요일(리스트 정보) entity 전환 및 저장
-        List<TblPatientTime> savePatientTimeList = new ArrayList<>();
-        for(PatientBaseDto.SavePatientTimeInfo ptDto : dto.getTimeList())
-            savePatientTimeList.add(ptDto.toEntity(patient));
-
-        patientTimeRepository.saveAll(savePatientTimeList);
-
-        //4. 어르신 정보 구분자 값, 담당자 이메일 반환
-        return PatientResponseDto.SavePatientInfo.builder()
-                .patientSeq(patient.getId())
-                .managerEmail(userEmail)
-                .build();
-    }
-
-    /**
-     * 어르신 정보 수정
-     * @param userEmail String:토큰 조회 결과 사용자 이메일 데이터
-     * @param dto PatientRequestDto.UpdatePatientInfo
-     * @return PatientResponseDto.UpdatePatientInfo
-     */
-    @Override
-    @Transactional
-    public PatientResponseDto.UpdatePatientInfo updatePatientInfo(String userEmail, PatientRequestDto.UpdatePatientInfo dto) {
-        //1. 관리자 정보 조회
-        TblCenterManager centerManager = loginAuthCheckService.checkManager(userEmail);
-
-        //2. 어르신 정보 조회 (어르신 구분자 & 관리자 구분자) 및 수정
-        //2-1. 어르신 데이터 조회
-        TblPatient patient = checkPatientManagerAuth(dto.getPatientSeq(), centerManager.getId());
-
-        //2-2. 주소 정보 검증
-        TblAddressFirst firstAddress = firstAddressService.validationCheck(dto.getAfSeq());
-        TblAddressSecond secondAddress = secondAddressService.validationCheck(dto.getAfSeq(), dto.getAsSeq());
-        TblAddressThird thirdAddress = thirdAddressService.validationCheck(dto.getAsSeq(), dto.getAtSeq());
-
-        //2-3. 데이터 수정
-        patient.basicUpdate(dto, firstAddress, secondAddress, thirdAddress);
-
-        //3. 어르신 돌봄 시간 요일 조회 및 삭제
-        List<TblPatientTime> patientTimeList = patientTimeRepository.findByPatient_Id(patient.getId());
-        patientTimeRepository.deleteAll(patientTimeList);
-
-        //4. 어르신 돌봄 시간 요일(리스트 정보) entity 전환 및 저장
-        List<TblPatientTime> savePatientTimeList = new ArrayList<>();
-        for(PatientBaseDto.SavePatientTimeInfo ptDto : dto.getTimeList())
-            savePatientTimeList.add(ptDto.toEntity(patient));
-
-        patientTimeRepository.saveAll(savePatientTimeList);
-
-        //5. 어르신 정보 구분자 값, 이름, 생년월일 중 연도 반환
-        return PatientResponseDto.UpdatePatientInfo.builder()
-                .patientSeq(patient.getId())
-                .managerEmail(userEmail)
-                .build();
-    }
-
-    /**
-     * 어르신 정보 상세 조회
-     * @param userEmail String:토큰 조회 결과 사용자 이메일 데이터
-     * @param patientSeq Long
-     * @return PatientResponseDto.GetOnePatientInfo
-     */
-    @Override
-    public PatientResponseDto.GetOnePatientInfo getOnePatientInfo(String userEmail, Long patientSeq) {
-        //1. 관리자 정보 조회
-        TblCenterManager centerManager = loginAuthCheckService.checkManager(userEmail);
-
-        //2. 어르신 정보 조회 (어르신 구분자 & 관리자 구분자)
-        TblPatient patient = checkPatientManagerAuth(patientSeq, centerManager.getId());
-
-        //3. 어르신 돌봄 시간 요일 조회
-        List<TblPatientTime> patientTimeList = patientTimeRepository.findByPatient_Id(patient.getId());
-
-        //4. 반환 dto 생성
-        PatientResponseDto.GetOnePatientInfo resultDto = new PatientResponseDto.GetOnePatientInfo(patient, patientTimeList);
-        resultDto.setCareChoice(commonService.getCareChoiceList(resultDto, false));
-        resultDto.setCareBaseDtoNull();
-        return resultDto;
-    }
-
-    @Override
-    public PatientResponseDto.GetPatientList getPatientList(String userEmail, PatientRequestDto.GetPatientList dto) {
-        //1. 관리자 정보 조회
-        TblCenterManager centerManager = loginAuthCheckService.checkManager(userEmail);
-
-        //2. 해당 관리자가 관리중인 어르신 리스트 조회
-        Pageable page = PageRequest.of(
-                dto.getPageNo()==null ? 0 : dto.getPageNo()
-                , dto.getPageSize()==null ? 10 : dto.getPageSize()
-                , Sort.by(Sort.Order.asc("name"), Sort.Order.desc("birthDate"), Sort.Order.desc("id")));
-        Page<TblPatient> listPage = patientRepository.findByManagerId(centerManager.getId(), page);
-        List<TblPatient> patientEntityList = listPage.getContent();
-
-        //3. 성별/근무종류/주소지/장기요양등급 각 항목 이름 조회 및 dto 리스트 정리
-        List<PatientResponseDto.GetPatientListInfo> resultList = new ArrayList<>();
-        for (TblPatient entity : patientEntityList) {
-            resultList.add(
-                    PatientResponseDto.GetPatientListInfo.builder()
-                            .patientSeq(entity.getId())
-                            .imgAddress(entity.getImg() == null ? null : fileService.getOneImgUrl(entity.getImg().getId()))
-                            .name(entity.getName())
-                            .age(codeService.calculateAge(LocalDate.parse(entity.getBirthDate(), java.time.format.DateTimeFormatter.BASIC_ISO_DATE)))
-                            .address(codeService.fullAddressString(entity.getTblAddressFirst(), entity.getTblAddressSecond(), entity.getTblAddressThird()))
-                            .genderStr(careRepository.findCalNameListStr(TblCareTopEnum.GENDER.getCareSeq(), entity.getGender()))
-                            .careLevelStr(careRepository.findCalNameListStr(TblCareTopEnum.CARE_LEVEL.getCareSeq(), entity.getCareLevel()))
-                            .workType(careRepository.findCalNameListStr(TblCareTopEnum.WORK_TYPE.getCareSeq(), entity.getInmateState()))
-                            .build());
-        }
-
-        //4. 결과 반환
-        return PatientResponseDto.GetPatientList.builder()
-                .totalPage(listPage.getTotalPages())
-                .totalEle(listPage.getTotalElements())
-                .hasNext(listPage.hasNext())
-                .list(resultList)
-                .build();
-    }
+//    /**
+//     * 어르신 정보 등록
+//     * @param userEmail String:토큰 조회 결과 사용자 이메일 데이터
+//     * @param dto PatientRequestDto.SavePatientInfo
+//     * @return PatientResponseDto.SavePatientInfo
+//     */
+//    @Override
+//    @Transactional
+//    public PatientResponseDto.SavePatientInfo savePatientInfo(String userEmail, PatientRequestDto.SavePatientInfo dto) {
+//        //1. 관리자 정보 조회
+//        TblCenterManager centerManager = loginAuthCheckService.checkManager(userEmail);
+//
+//        //2. 어르신 정보 등록
+//        //2-1. 주소 정보 검증
+//        TblAddressFirst firstAddress = firstAddressService.validationCheck(dto.getAfSeq());
+//        TblAddressSecond secondAddress = secondAddressService.validationCheck(dto.getAfSeq(), dto.getAsSeq());
+//        TblAddressThird thirdAddress = thirdAddressService.validationCheck(dto.getAsSeq(), dto.getAtSeq());
+//
+//        //2-2. 어르신 정보 entity 전환 및 저장
+//        TblPatient patient = patientRepository.save(dto.toEntity(centerManager, firstAddress, secondAddress, thirdAddress));
+//
+//        //3. 어르신 돌봄 시간 요일(리스트 정보) entity 전환 및 저장
+//        List<TblPatientTime> savePatientTimeList = new ArrayList<>();
+//        for(PatientBaseDto.SavePatientTimeInfo ptDto : dto.getTimeList())
+//            savePatientTimeList.add(ptDto.toEntity(patient));
+//
+//        patientTimeRepository.saveAll(savePatientTimeList);
+//
+//        //4. 어르신 정보 구분자 값, 담당자 이메일 반환
+//        return PatientResponseDto.SavePatientInfo.builder()
+//                .patientSeq(patient.getId())
+//                .managerEmail(userEmail)
+//                .build();
+//    }
+//
+//    /**
+//     * 어르신 정보 수정
+//     * @param userEmail String:토큰 조회 결과 사용자 이메일 데이터
+//     * @param dto PatientRequestDto.UpdatePatientInfo
+//     * @return PatientResponseDto.UpdatePatientInfo
+//     */
+//    @Override
+//    @Transactional
+//    public PatientResponseDto.UpdatePatientInfo updatePatientInfo(String userEmail, PatientRequestDto.UpdatePatientInfo dto) {
+//        //1. 관리자 정보 조회
+//        TblCenterManager centerManager = loginAuthCheckService.checkManager(userEmail);
+//
+//        //2. 어르신 정보 조회 (어르신 구분자 & 관리자 구분자) 및 수정
+//        //2-1. 어르신 데이터 조회
+//        TblPatient patient = checkPatientManagerAuth(dto.getPatientSeq(), centerManager.getId());
+//
+//        //2-2. 주소 정보 검증
+//        TblAddressFirst firstAddress = firstAddressService.validationCheck(dto.getAfSeq());
+//        TblAddressSecond secondAddress = secondAddressService.validationCheck(dto.getAfSeq(), dto.getAsSeq());
+//        TblAddressThird thirdAddress = thirdAddressService.validationCheck(dto.getAsSeq(), dto.getAtSeq());
+//
+//        //2-3. 데이터 수정
+//        patient.basicUpdate(dto, firstAddress, secondAddress, thirdAddress);
+//
+//        //3. 어르신 돌봄 시간 요일 조회 및 삭제
+//        List<TblPatientTime> patientTimeList = patientTimeRepository.findByPatient_Id(patient.getId());
+//        patientTimeRepository.deleteAll(patientTimeList);
+//
+//        //4. 어르신 돌봄 시간 요일(리스트 정보) entity 전환 및 저장
+//        List<TblPatientTime> savePatientTimeList = new ArrayList<>();
+//        for(PatientBaseDto.SavePatientTimeInfo ptDto : dto.getTimeList())
+//            savePatientTimeList.add(ptDto.toEntity(patient));
+//
+//        patientTimeRepository.saveAll(savePatientTimeList);
+//
+//        //5. 어르신 정보 구분자 값, 이름, 생년월일 중 연도 반환
+//        return PatientResponseDto.UpdatePatientInfo.builder()
+//                .patientSeq(patient.getId())
+//                .managerEmail(userEmail)
+//                .build();
+//    }
+//
+//    /**
+//     * 어르신 정보 상세 조회
+//     * @param userEmail String:토큰 조회 결과 사용자 이메일 데이터
+//     * @param patientSeq Long
+//     * @return PatientResponseDto.GetOnePatientInfo
+//     */
+//    @Override
+//    public PatientResponseDto.GetOnePatientInfo getOnePatientInfo(String userEmail, Long patientSeq) {
+//        //1. 관리자 정보 조회
+//        TblCenterManager centerManager = loginAuthCheckService.checkManager(userEmail);
+//
+//        //2. 어르신 정보 조회 (어르신 구분자 & 관리자 구분자)
+//        TblPatient patient = checkPatientManagerAuth(patientSeq, centerManager.getId());
+//
+//        //3. 어르신 돌봄 시간 요일 조회
+//        List<TblPatientTime> patientTimeList = patientTimeRepository.findByPatient_Id(patient.getId());
+//
+//        //4. 반환 dto 생성
+//        PatientResponseDto.GetOnePatientInfo resultDto = new PatientResponseDto.GetOnePatientInfo(patient, patientTimeList);
+//        resultDto.setCareChoice(commonService.getCareChoiceList(resultDto, false));
+//        resultDto.setCareBaseDtoNull();
+//        return resultDto;
+//    }
+//
+//    @Override
+//    public PatientResponseDto.GetPatientList getPatientList(String userEmail, PatientRequestDto.GetPatientList dto) {
+//        //1. 관리자 정보 조회
+//        TblCenterManager centerManager = loginAuthCheckService.checkManager(userEmail);
+//
+//        //2. 해당 관리자가 관리중인 어르신 리스트 조회
+//        Pageable page = PageRequest.of(
+//                dto.getPageNo()==null ? 0 : dto.getPageNo()
+//                , dto.getPageSize()==null ? 10 : dto.getPageSize()
+//                , Sort.by(Sort.Order.asc("name"), Sort.Order.desc("birthDate"), Sort.Order.desc("id")));
+//        Page<TblPatient> listPage = patientRepository.findByManagerId(centerManager.getId(), page);
+//        List<TblPatient> patientEntityList = listPage.getContent();
+//
+//        //3. 성별/근무종류/주소지/장기요양등급 각 항목 이름 조회 및 dto 리스트 정리
+//        List<PatientResponseDto.GetPatientListInfo> resultList = new ArrayList<>();
+//        for (TblPatient entity : patientEntityList) {
+//            resultList.add(
+//                    PatientResponseDto.GetPatientListInfo.builder()
+//                            .patientSeq(entity.getId())
+//                            .imgAddress(entity.getImg() == null ? null : fileService.getOneImgUrl(entity.getImg().getId()))
+//                            .name(entity.getName())
+//                            .age(codeService.calculateAge(LocalDate.parse(entity.getBirthDate(), java.time.format.DateTimeFormatter.BASIC_ISO_DATE)))
+//                            .address(codeService.fullAddressString(entity.getTblAddressFirst(), entity.getTblAddressSecond(), entity.getTblAddressThird()))
+//                            .genderStr(careRepository.findCalNameListStr(TblCareTopEnum.GENDER.getCareSeq(), entity.getGender()))
+//                            .careLevelStr(careRepository.findCalNameListStr(TblCareTopEnum.CARE_LEVEL.getCareSeq(), entity.getCareLevel()))
+//                            .workType(careRepository.findCalNameListStr(TblCareTopEnum.WORK_TYPE.getCareSeq(), entity.getInmateState()))
+//                            .build());
+//        }
+//
+//        //4. 결과 반환
+//        return PatientResponseDto.GetPatientList.builder()
+//                .totalPage(listPage.getTotalPages())
+//                .totalEle(listPage.getTotalElements())
+//                .hasNext(listPage.hasNext())
+//                .list(resultList)
+//                .build();
+//    }
 
     /**
      * 어르신 공고 등록
